+++ core/src/components/icon/bn-icon.tsx
@@ -1,15 +1,5 @@
-import {
-  Component,
-  ComponentInterface,
-  Event,
-  EventEmitter,
-  getAssetPath,
-  h,
-  Host,
-  Prop,
-  State,
-  Watch,
-} from "@stencil/core";
+import { Component, ComponentInterface, Event, EventEmitter, h, Host, Prop, State, Watch } from "@stencil/core";
+import { assets } from "../../utils/assets";
 import bnIconMetadata from "./assets/bnqkl-icon.json";
 import type { $Direction } from "./bn-icon.const";
 import type { $BnIconName } from "./bn-icon.name";
@@ -30,7 +20,7 @@ export class BnIcon implements ComponentInterface {
       gloablStyle.innerHTML = `
       @font-face {
         font-family: "bnqkl-icon";
-        src: url("${getAssetPath("./bnqkl-icon.woff")}") format("woff");
+        src: url("${assets.get("./bnqkl-icon.woff")}") format("woff");
         font-weight: normal;
         font-style: normal;
         font-display: block;