+++ core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bnqkl/web-component
 
+## 1.1.1
+
+### Patch Changes
+
+- fix assets path
+
+## 1.1.0
+
+### Minor Changes
+
+- fix icon asset path
+
 ## 1.0.2
 
 ### Patch Changes